🗺️ Atlas: Refactor temporal index structure
🕸️ Tangle: `src/index/temporal.rs` was a 4000-line "Blob" mixing configuration, policy, data structures, and core logic. It also had a name collision for `VersionMetadata`.

📐 Blueprint:
- Extracted `TemporalIndexConfig` to `src/index/temporal/config.rs`.
- Extracted `DeduplicationPolicy` to `src/index/temporal/policy.rs`.
- Extracted `TimelineVersionMetadata` (renamed from `VersionMetadata`) to `src/index/temporal/metadata.rs`.
- Extracted `EntityTimeline` and `TimelineEntry` to `src/index/temporal/timeline.rs`.
- Kept `TemporalIndexes` in `src/index/temporal/mod.rs` as the facade.
- Moved tests to `src/index/temporal/tests.rs`.

🧱 Stability: Preserved all public APIs. Resolved `VersionMetadata` name collision.

🔬 Verification: `cargo test index::temporal` passes.
